requirements.txt
anyio==3.6.1
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
fastapi==0.85.0
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.1
numpy==1.23.3
pandas==1.5.0
pydantic==1.10.2
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
